move registrants routes to own folder
