+++ apps/api/src/app.module.ts
@@ -7,9 +7,17 @@ import { AppLogger } from './common/logger.service';
 import { UserModule } from './user/user.module';
 import { AuthModule } from './auth/auth.module';
 import { AdminModule } from './admin/admin.module';
+import { RegistrantModule } from './registrant/registrant.module';
 
 @Module({
-  imports: [AuthModule, DatabaseModule, SubmissionModule, UserModule, AdminModule],
+  imports: [
+    AuthModule,
+    DatabaseModule,
+    SubmissionModule,
+    UserModule,
+    AdminModule,
+    RegistrantModule,
+  ],
   controllers: [AppController],
   providers: [AppService, AppLogger],
 })
