+++ apps/api/src/submission/submission.service.ts
@@ -7,7 +7,6 @@ import {
   PersonalInformationDTO,
   UpdateSubmissionDTO,
   SubmissionRO,
-  RegistrantRO,
   RegistrantFilterDTO,
 } from '@ehpr/common';
 import { MailService } from 'src/mail/mail.service';
@@ -16,7 +15,6 @@ import { Recipient } from 'src/mail/types/recipient';
 import { generateConfirmationId } from './id-generator';
 import { AppLogger } from 'src/common/logger.service';
 import { UpdateConfirmationMailable } from 'src/mail/mailables/update-confirmation.mailable';
-import { formatRegistrants } from './submission.util';
 
 @Injectable()
 export class SubmissionService {
@@ -77,15 +75,6 @@ export class SubmissionService {
 
   async getSubmissions() {
     return this.submissionRepository.find();
-  }
-
-  async getRegistrants(
-    filter: RegistrantFilterDTO,
-  ): Promise<[data: RegistrantRO[], count: number]> {
-    const [data, count] = await this.getSubmissionsFilterQuery(filter);
-
-    const registrants = formatRegistrants(data);
-    return [registrants, count];
   }
 
   async getSubmissionsFilterQuery(filter: RegistrantFilterDTO) {
