+++ apps/web/src/services/index.ts
@@ -1,3 +1,4 @@
 export * from './admin';
 export * from './submission';
 export * from './user';
+export * from './registrant';
